Refactored linear convolution code and added alternative convolution methods. Also fixed data type issue in unit impulse code.

+++ waves/unit_impulse.py
@@ -1,7 +1,7 @@
 import numpy as np
 import matplotlib.pyplot as plt
 x= np.arange(-10,10,0.00001)
-y=np.zeros(len(x),dtype=int)
+y=np.zeros(len(x))
 y[np.isclose(x,0,atol=0.000001)]=1
 plt.stem(x,y)
 plt.show()
